Web Rename the MediaViewer file to index
